# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Request API",
        "description": "Request API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/requests": {
            "get": {
                "tags": [
                    "Requests"
                ],
                "summary": "Return requests list with pagination",
                "description": "Full requests list",
                "operationId": "85eff1e5d0c6e43d8063ccc4db4dfa8a",
                "responses": {
                    "default": {
                        "description": "Requests list",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "api/requests/": {
            "put": {
                "tags": [
                    "Requests"
                ],
                "summary": "Resolve active request with moderator's comment",
                "operationId": "resolveRequest",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Requests"
                ],
                "summary": "Store new request",
                "operationId": "storeRequest",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        }
    }
}